+++ compute/api/src/lib.rs
@@ -3,10 +3,12 @@ extern crate grpcio;
 extern crate protobuf;
 
 extern crate ekiden_common_api;
+extern crate ekiden_consensus_api;
 
 mod generated;
 
 use ekiden_common_api as common;
+use ekiden_consensus_api as consensus;
 
 pub use generated::computation_group::*;
 pub use generated::computation_group_grpc::*;